Factor out memory allocation for conversion from baseXX
After using valgrind along with vgdb, the reason why calloc(3) is
required to avoid warnings became clear: in byte_from_baseXX the value
of cc (the partial value of a whole byte) is logically or'ed with memory
which -- if not properly zeroed -- is uninitialized. One could zero a
new byte during conversion, but using calloc(3) does the same thing less
complicated and most likeley way faster.

Signed-off-by: Thomas Egerer <thomas.egerer@port22.eu>